+++ emptylog/__init__.py
@@ -1,5 +1,5 @@
 from emptylog.empty_logger import EmptyLogger as EmptyLogger  # noqa: F401
-from emptylog.protocol import LoggerProtocol as LoggerProtocol  # noqa: F401
+from emptylog.protocols import LoggerProtocol as LoggerProtocol  # noqa: F401
 from emptylog.memory_logger import MemoryLogger as MemoryLogger  # noqa: F401
 from emptylog.printing_logger import PrintingLogger as PrintingLogger  # noqa: F401
 from emptylog.loggers_group import LoggersGroup as LoggersGroup  # noqa: F401
