+++ emptylog/protocols.py
